requirements.txt
google-adk>=0.3.0
litellm>=1.40.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pandas>=2.0.0
scikit-learn>=1.4.0
opencv-python>=4.8.0
httpx>=0.25.0
chromadb>=0.5.0
plotly>=5.20.0
ollama>=0.1.7
openpyxl>=3.1.2
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
